requirements.txt
# Core dependencies
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.5.0
python-dotenv>=1.0.0

# LangChain and AI
langchain>=0.1.0
langchain-anthropic>=0.1.0
langgraph>=0.2.6

# Pinecone
pinecone>=3.0.0

# Telemetry and logging
python-json-logger>=2.0.0

# Testing
pytest>=8.0.0
pytest-asyncio>=0.21.0

# Development
ruff>=0.1.0
mypy>=1.7.0
